refactor to new architecture

+++ scripts/install-task.sh
@@ -1,7 +1,13 @@
 #!/usr/bin/env bash
 
-curl -sL https://taskfile.dev/install.sh | sh
-mv bin/task $HOME/.local/bin/
-echo "task binary moved to $HOME/.local/bin"
-echo "please do 'source ~/.bashrc' to reload \$PATH"
-rm -R bin
+TASKPATH=$(which task)
+if [ -z "$TASKPATH" ]
+then
+  curl -sL https://taskfile.dev/install.sh | sh
+  mv bin/task $HOME/.local/bin/
+  echo "task binary moved to $HOME/.local/bin"
+  echo "please do 'source ~/.bashrc' to reload \$PATH"
+  rm -R bin
+else
+  echo "Task has already been installed"
+fi
